+++ fx-domain/src/main/java/fx/domain/instruments.java
@@ -1,0 +1,5 @@
+package fx.domain;
+
+public class Instruments {
+    public enum Instrument {EUR_USD, USD_EUR}
+}
